requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
scikit-learn
statsmodels
arch
xgboost
lightgbm
tensorflow
torch
yfinance
streamlit
plotly
pyyaml